Stabilize paginated data flows and real-data end-to-end coverage
The app had regressed after the paginated data migration, and the automated verification surfaces were too brittle to prove the fixes. This change completes the data-path repair, aligns the in-memory/test backends with pagination-first behavior, and hardens the native and web end-to-end suites so they can run against real source data and current responsive layouts.

Constraint: Existing look4k source data must be exercised through local ignored credentials, not committed fixtures
Constraint: Web test backend could not use port 8080 because it conflicts with another local service
Rejected: Keep relying on legacy bulk VOD state in Home and tests | masked the real production data path
Rejected: Leave Playwright on its default bundled browser/runtime setup | failed on this host and blocked reproducible web verification
Confidence: high
Scope-risk: moderate
Directive: Keep Home/VOD verification on paginated providers; do not reintroduce bulk-state shortcuts for tests
Tested: flutter test integration_test/main_test.dart -d linux
Tested: flutter test test/features/home/presentation/providers/home_providers_test.dart
Tested: Playwright smoke/navigation/responsive/search/settings/live_tv/vod/favorites/dvr-schedule/player/source-sync/comprehensive-crawl against web build + crispy-server on port 8081
Not-tested: Android emulator execution with the newly prepared AVDs
Reversibility: clean

+++ lib/core/data/ffi_backend_channels.dart
@@ -64,7 +64,7 @@ mixin _FfiChannelsMixin on _FfiBackendBase {
       sourceIdsJson: sourceIdsJson,
       group: group,
     );
-    return result;
+    return result.toInt();
   }
 
   Future<List<String>> getChannelIdsForGroup(
